[FLINK-17786][sql-client] Remove dialect from ExecutionEntry
